I did not commit after each section so adding a very general commit message for all changes after Hello, World up to half-way through Arrays and slice.
Added all code for the Integers section of Learn Go with Tests. This included code related to adding integers and code to test them.

Added all code for the integers section of Learn Go with Tests. This included code related to for loops and code test them.

Added all code from the start of the Arrays and slices section up to the inclusion of the SumAll function example. This includes code related to slices and code to test them.